% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenograph.R
\name{find_neighbors}
\alias{find_neighbors}
\title{K Nearest Neighbour Search}
\usage{
find_neighbors(data, k)
}
\arguments{
\item{data}{matrix; input data matrix}

\item{k}{integer; number of nearest neighbours}
}
\value{
a n-by-k matrix of neighbor indices
}
\description{
K Nearest Neighbour Search
}
\examples{
iris_unique <- unique(iris) # Remove duplicates
data <- as.matrix(iris_unique[,1:4])
neighbors <- find_neighbors(data, k=10)

}
