% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenograph.R
\name{find_neighbors_rand_proj}
\alias{find_neighbors_rand_proj}
\title{Fast approximate KNN using randomProjectionTreeSearch
as implemented in largeVis}
\usage{
find_neighbors_rand_proj(data, k, ...)
}
\arguments{
\item{data}{matrix; input data matrix}

\item{k}{integer; number of nearest neighbours}
}
\value{
a n-by-k matrix of neighbor indices
}
\description{
Fast approximate KNN using randomProjectionTreeSearch
as implemented in largeVis
}
\examples{
iris_unique <- unique(iris) # Remove duplicates
data <- as.matrix(iris_unique[,1:4])
neighbors <- find_neighbors_rand_proj(t(data), k=10)

}
